+++ js/dterm.js
@@ -12,7 +12,17 @@
  */
 /* global define */
 (function(factory, undefined) {
-    var root = typeof window !== 'undefined' ? window : global;
+
+    var root;
+    if (typeof window !== 'undefined') {
+        root = window;
+    } else if (typeof self !== 'undefined') {
+        root = self;
+    } else if (typeof global !== 'undefined') {
+        root = global;
+    } else {
+        throw new Error('Unknow context');
+    }
     if (typeof define === 'function' && define.amd) {
         // AMD. Register as an anonymous module.
         // istanbul ignore next